refactor: change location navbar file
